WIP fixing map view toolbar
